requirements.txt
# ===== FRAMEWORK WEB =====
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# ===== FIREBASE =====
firebase-admin==6.2.0

# ===== GOOGLE APIS (Gmail) =====
google-api-python-client==2.101.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# ===== VALIDACIÓN Y CONFIGURACIÓN =====
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# ===== SEGURIDAD =====
bcrypt==4.1.1
PyJWT==2.10.1

# ===== PLANTILLAS Y EMAIL =====
jinja2==3.1.2

# ===== LOGGING =====
python-json-logger==2.0.7

# ===== DESARROLLO (OPCIONAL) =====
pytest==7.4.3
httpx==0.25.2
